+++ cryosparc/command.py
@@ -4,12 +4,20 @@ servers. Generally should not be used directly.
 """
 from contextlib import contextmanager
 import json
+import os
 import socket
+import time
 import uuid
 from typing import Optional, Type
 from urllib.request import urlopen, Request
 from urllib.error import HTTPError, URLError
 from urllib.parse import urlencode
+from warnings import warn
+
+from .errors import CommandError
+
+MAX_ATTEMPTS = int(os.getenv("CRYOSPARC_COMMAND_RETRIES", 3))
+RETRY_INTERVAL = int(os.getenv("CRYOSPARC_COMMAND_RETRY_SECONDS", 30))
 
 
 class CommandClient:
@@ -67,12 +75,8 @@ class CommandClient:
 
     """
 
+    Error = CommandError
     service: str
-
-    class Error(Exception):
-        def __init__(self, parent: "CommandClient", reason: str, *args: object, url: str = "") -> None:
-            msg = f"*** {type(parent).__name__}: ({url}) {reason}"
-            super().__init__(msg, *args)
 
     def __init__(
         self,
@@ -94,24 +98,35 @@ class CommandClient:
     def _get_callable(self, key):
         def func(*args, **kwargs):
             params = kwargs if len(kwargs) else args
-            data = {
-                "jsonrpc": "2.0",
-                "method": key,
-                "params": params,
-                "id": str(uuid.uuid4()),
-            }
+            data = {"jsonrpc": "2.0", "method": key, "params": params, "id": str(uuid.uuid4())}
             res = None
             try:
-                with make_json_request(self, "/api", data=data) as request:
+                with make_json_request(self, "/api", data=data, _stacklevel=4) as request:
                     res = json.loads(request.read())
-            except CommandClient.Error as err:
-                raise CommandClient.Error(
-                    self, f'Did not receive a JSON response from method "{key}" with params {params}', url=self._url
+            except CommandError as err:
+                raise CommandError(
+                    f'Encounted error from JSONRPC function "{key}" with params {params}',
+                    url=self._url,
+                    code=err.code,
+                    data=err.data,
                 ) from err
 
-            assert res, f'JSON response not received for method "{key}" with params {params}'
-            assert "error" not in res, f'Error for "{key}" with params {params}:\n' + format_server_error(res["error"])
-            return res["result"]
+            if not res:
+                raise CommandError(
+                    f'JSON response not received from JSONRPC function "{key}" with params {params}',
+                    url=self._url,
+                )
+            elif "error" in res:
+                error = res["error"]
+                raise CommandError(
+                    f'Encountered {error.get("name", "Error")} from JSONRPC function "{key}" with params {params}:\n'
+                    f"{format_server_error(error)}",
+                    url=self._url,
+                    code=error.get("code"),
+                    data=error.get("data"),
+                )
+            else:
+                return res["result"]  # OK
 
         return func
 
@@ -127,7 +142,14 @@ class CommandClient:
 
 @contextmanager
 def make_request(
-    client: CommandClient, method: str = "POST", url: str = "", query: dict = {}, data=None, headers: dict = {}
+    client: CommandClient,
+    method: str = "POST",
+    url: str = "",
+    *,
+    query: dict = {},
+    data=None,
+    headers: dict = {},
+    _stacklevel=2,  # controls warning line number
 ):
     """
     Create a raw HTTP request/response context with the given command client.
@@ -141,7 +163,7 @@ def make_request(
         headers (dict, optional): HTTP headers. Defaults to {}.
 
     Raises:
-        CommandClient.Error: General error such as timeout, URL or HTTP
+        CommandError: General error such as timeout, URL or HTTP
 
     Yields:
         http.client.HTTPResponse:  Use with a context manager to get HTTP response
@@ -157,42 +179,53 @@ def make_request(
     url = f"{client._url}{url}{'?' + urlencode(query) if query else ''}"
     headers = {"Originator": "client", **client._headers, **headers}
     attempt = 1
-    max_attempts = 3
     error_reason = "<unknown>"
-    while attempt < max_attempts:
+    code = 500
+    resdata = None
+    while attempt < MAX_ATTEMPTS:
         request = Request(url, data=data, headers=headers, method=method)
         response = None
         try:
             with urlopen(request, timeout=client._timeout) as response:
                 yield response
                 return
-        except (TimeoutError, socket.timeout):
-            error_reason = "Timeout Error"
-            print(
-                f"*** {type(client).__name__}: command ({url}) "
-                f"did not reply within timeout of {client._timeout} seconds, "
-                f"attempt {attempt} of {max_attempts}"
-            )
-            attempt += 1
-        except HTTPError as error:
+        except HTTPError as error:  # command server reported an error
+            code = error.code
             error_reason = (
                 f"HTTP Error {error.code} {error.reason}; "
                 f"please check cryosparcm log {client.service} for additional information."
             )
             if error.readable():
-                error_reason += "\nResponse from server: " + str(error.read())
+                resdata = error.read()
+                error_reason += f"\nResponse from server: {data}"
+            if resdata and error.headers.get_content_type() == "application/json":
+                resdata = json.loads(resdata)
 
-            print(f"*** {type(client).__name__}: ({url}) {error_reason}")
+            warn(f"*** {type(client).__name__}: ({url}) {error_reason}", stacklevel=_stacklevel)
             break
-        except URLError as error:
+        except URLError as error:  # command server may be down
             error_reason = f"URL Error {error.reason}"
-            print(f"*** {type(client).__name__}: ({url}) {error_reason}")
-            break
+            warn(
+                f"*** {type(client).__name__}: ({url}) {error_reason}, attempt {attempt} of {MAX_ATTEMPTS}. "
+                f"Retrying in {RETRY_INTERVAL} seconds",
+                stacklevel=_stacklevel,
+            )
+            time.sleep(RETRY_INTERVAL)
+            attempt += 1
+        except (TimeoutError, socket.timeout):  # slow network connection or request
+            error_reason = "Timeout Error"
+            warn(
+                f"*** {type(client).__name__}: command ({url}) "
+                f"did not reply within timeout of {client._timeout} seconds, "
+                f"attempt {attempt} of {MAX_ATTEMPTS}",
+                stacklevel=_stacklevel,
+            )
+            attempt += 1
 
-    raise CommandClient.Error(client, error_reason, url=url)
+    raise CommandError(error_reason, url=url, code=code, data=resdata)
 
 
-def make_json_request(client: CommandClient, url="", query={}, data=None, headers={}):
+def make_json_request(client: CommandClient, url="", *, query={}, data=None, headers={}, _stacklevel=3):
     """
     Similar to ``make_request``, except sends request body data JSON and
     receives arbitrary response.
@@ -208,6 +241,9 @@ def make_json_request(client: CommandClient, url="", query={}, data=None, header
     Yields:
         http.client.HTTPResponse: Use with a context manager to get HTTP response
 
+    Raises:
+        CommandError: General error such as timeout, URL or HTTP
+
     Example:
 
         >>> from cryosparc.command import CommandClient, make_json_request
@@ -218,10 +254,13 @@ def make_json_request(client: CommandClient, url="", query={}, data=None, header
     """
     headers = {"Content-Type": "application/json", **headers}
     data = json.dumps(data, cls=client._cls).encode()
-    return make_request(client, url=url, query=query, data=data, headers=headers)
+    return make_request(client, url=url, query=query, data=data, headers=headers, _stacklevel=_stacklevel)
 
 
 def format_server_error(error):
+    """
+    :meta private:
+    """
     err = error["message"] if "message" in error else str(error)
     if "data" in error and error["data"]:
         if isinstance(error["data"], dict) and "traceback" in error["data"]: