+++ cryosparc/core.pyx
@@ -1,7 +1,10 @@
 from . cimport dataset
-from libc.stdint cimport uint8_t, uint32_t
-from cython.view cimport array
+from . cimport lz4
+from libc.stdint cimport uint8_t, uint32_t, uint64_t
 from cpython.ref cimport PyObject, Py_XINCREF, Py_XDECREF
+from cpython.mem cimport PyMem_Realloc, PyMem_Free
+
+cdef int LZ4_ACCELERATION = 1000
 
 
 # Mirror of equivalent C-datatype enumeration
@@ -24,7 +27,6 @@ cpdef enum DsetType:
 
 cdef class Data:
     cdef dataset.Dset _handle
-    cdef dict _strcache
 
     def __cinit__(self, other = None):
         cdef Data othr
@@ -42,8 +44,6 @@ cdef class Data:
 
         if self._handle == 0:
             raise MemoryError()
-
-        self._strcache = dict()
 
     def __dealloc__(self):
         if self._handle:
@@ -128,7 +128,7 @@ cdef class Data:
 
     def getbuf(self, str colkey):
         cdef void *mem
-        cdef Py_ssize_t size
+        cdef size_t size
         cdef bytes colkey_b = colkey.encode()
         cdef const char *colkey_c = colkey_b
         with nogil:
@@ -139,8 +139,8 @@ cdef class Data:
         else:
             return <unsigned char [:size]> mem
 
-    def getstr(self, str colkey, Py_ssize_t index):
-        return dataset.dset_getstr(self._handle, colkey.encode(), index)  # returns bytes
+    def getstr(self, str col, size_t index):
+        return dataset.dset_getstr(self._handle, col.encode(), index)  # returns bytes
 
     def tocstrs(self, str col):
         # Convert Python strings to C strings in the given object column
@@ -160,7 +160,7 @@ cdef class Data:
             pybytes = pystr.encode()
             Py_XDECREF(pycol[i])  # so string is deallocated
             pycol[i] = NULL  # so that strfree not attempted
-            dataset.dset_setstr(self._handle, colkey, i, pybytes)
+            dataset.dset_setstr(self._handle, colkey, i, pybytes, len(pybytes))
 
         return True
 
@@ -192,11 +192,241 @@ cdef class Data:
 
         return bool(dataset.dset_changecol(self._handle, colkey, T_OBJ))
 
+    def stralloc(self, str val):
+        cdef uint64_t idx
+        cdef bytes pybytes = val.encode()
+        if not dataset.dset_stralloc(self._handle, pybytes, len(pybytes), &idx):
+            raise MemoryError()
+        return <int> idx
+
+    def dump(self):
+        cdef void *mem
+        cdef size_t size
+        with nogil:
+            mem = dataset.dset_dump(self._handle)
+            size = dataset.dset_totalsz(self._handle)
+        return <unsigned char [:size]> mem
+
+    def dumpstrheap(self):
+        cdef void *mem
+        cdef size_t size
+        with nogil:
+            mem = dataset.dset_strheap(self._handle)
+            size = dataset.dset_strheapsz(self._handle)
+        return <unsigned char [:size]> mem
+
+    def setstrheap(self, bytes heap):
+        if not dataset.dset_setstrheap(self._handle, <const char *> heap, len(heap)):
+            raise MemoryError()
+
     def defrag(self, bint realloc_smaller):
         return dataset.dset_defrag(self._handle, realloc_smaller)
 
-    def dumptxt(self):
-        dataset.dset_dumptxt(self._handle)
+    def dumptxt(self, bint dump_data = 0):
+        dataset.dset_dumptxt(self._handle, dump_data)
 
     def handle(self):
         return self._handle
+
+
+cdef class Stream:
+    # Helper class for initializing a dataset from a compressed stream or
+    # generating a compressed stream for that dataset.
+    #
+    # WARNING: Methods here recycle result buffers to minimize allocations. If
+    # results are not consumed prior to future calls to this class, the contents
+    # will be overwritten.
+    cdef Data data
+    cdef char *buf  # internal compression buffer
+    cdef uint64_t *aux  # internal string index array
+    cdef size_t bufsz
+    cdef size_t auxsz
+
+    def __cinit__(self, Data data = None):
+        self.data = data
+        self.buf = NULL
+        self.aux = NULL
+        self.bufsz = 0
+        self.auxsz = 0
+
+    def __dealloc__(self):
+        if self.buf != NULL:
+            PyMem_Free(self.buf)
+            self.buf = NULL
+        if self.aux != NULL:
+            PyMem_Free(self.aux)
+            self.aux = NULL
+
+    def _ensure_buf(self, int min_len):
+        cdef size_t sz = min_len
+        if self.bufsz < sz:
+            self.buf = <char *> PyMem_Realloc(<void *> self.buf, sz)
+            self.bufsz = sz
+        return <int> self.bufsz
+
+    def _ensure_aux(self, int min_sz):
+        cdef size_t sz = min_sz
+        if self.auxsz < sz:
+            self.aux = <uint64_t *> PyMem_Realloc(<void *> self.aux, sz)
+            self.auxsz = sz
+        return <int> self.auxsz
+
+    def cast_objs_to_strs(self):
+        # change all T_OBJ column types to T_STR (underlying data unaffected)
+        cdef dataset.Dset handle = self.data._handle
+        cdef int ncol = self.data.ncol()
+        cdef int coltype
+        cdef const char *colkey
+
+        with nogil:
+            for i in xrange(ncol):
+                colkey = dataset.dset_key(handle, i)
+                coltype = dataset.dset_type(handle, colkey)
+                if coltype == T_OBJ:
+                    dataset.dset_changecol(handle, colkey, T_STR)
+
+    def stralloc_col(self, str col):
+        # Allocate C strings for every Python string in the given object column
+        # Returns 0 if the column does not have object type.
+        cdef dataset.Dset handle = self.data._handle
+        cdef bytes colkey = col.encode()
+        cdef int coltype = dataset.dset_type(handle, colkey)
+        if coltype != T_OBJ:
+            return 0  # invalid column
+
+        # Convert to C strings and compress index array instead
+        cdef PyObject **coldata = <PyObject **> dataset.dset_get(handle, colkey)
+        cdef uint64_t nrow = dataset.dset_nrow(handle)
+        cdef size_t sz = dataset.dset_getsz(handle, colkey)
+
+        self._ensure_aux(sz)
+
+        cdef uint64_t idx
+        cdef int allocres
+        cdef str pystr
+        cdef bytes pybytes
+
+        for i in xrange(nrow):
+            pystr = <str> coldata[i]
+            pybytes = pystr.encode()
+            allocres = dataset.dset_stralloc(handle, pybytes, len(pybytes), &idx)
+            if allocres == 0:
+                raise MemoryError()
+            elif allocres == 2:
+                # dataset reallocated, coldata must be retrieved
+                coldata = <PyObject **> dataset.dset_get(handle, colkey)
+            self.aux[i] = idx
+
+        return <unsigned char [:sz]> (<unsigned char *> self.aux)
+
+    def compress_col(self, str col):
+        # Use when multiple compression calls are required to minimize total
+        # number of allocations.
+        #
+        # Overwrites (and potentially re-allocates) internal buffer each time
+        # it's called so only the latest return value is valid memory (until the
+        # instance is garbage collected).
+        #
+        # If the column is has type T_OBJ, allocates each python string as
+        # a C string in the string heap and compresses the resulting array of
+        # indexes into the C string.
+        #
+        # Returns array memoryview with compressed data
+        cdef dataset.Dset handle = self.data._handle
+        cdef colkey = col.encode()
+        cdef int coltype = dataset.dset_type(handle, colkey)
+
+        if coltype == 0:
+            return 0  # invalid column
+
+        cdef uint64_t nrow = dataset.dset_nrow(handle)
+        cdef size_t sz = dataset.dset_getsz(handle, colkey)
+        cdef uint64_t idx
+        cdef int allocres
+        cdef str pystr
+        cdef bytes pybytes
+        cdef PyObject **coldata
+        cdef unsigned char [:] data
+
+        if coltype == T_OBJ:
+            # Convert to C strings and compress index array instead
+            self._ensure_aux(sz)
+            coldata = <PyObject **> dataset.dset_get(handle, colkey)
+
+            for i in xrange(nrow):
+                pystr = <str> coldata[i]
+                pybytes = pystr.encode()
+                allocres = dataset.dset_stralloc(handle, pybytes, len(pybytes), &idx)
+                if allocres == 0:
+                    raise MemoryError()
+                elif allocres == 2:
+                    # dataset reallocated, coldata must be retrieved
+                    coldata = <PyObject **> dataset.dset_get(handle, colkey)
+                self.aux[i] = idx
+
+            data = <unsigned char [:sz]> (<unsigned char *> self.aux)
+        else:
+            data = <unsigned char [:sz]> (<unsigned char *> dataset.dset_get(handle, colkey))
+
+        return self.compress(data)
+
+    def compress_numpy(self, arr):
+        cdef size_t sz = arr.size * arr.itemsize
+        cdef size_t arr_ptr_val = arr.ctypes.data
+        cdef void *arr_ptr = <void *> arr_ptr_val
+        return self.compress(<unsigned char [:sz]> arr_ptr)
+
+    def compress(self, unsigned char [:] data):
+        cdef int sz = data.size
+        cdef int max_compressed_sz = lz4.LZ4_compressBound(sz)
+
+        self._ensure_buf(max_compressed_sz)
+
+         # Call compress. Write final length into compressed_sz
+        cdef int compressed_sz = lz4.LZ4_compress_fast(
+            <const char *> &data[0],
+            self.buf,
+            sz,
+            max_compressed_sz,
+            LZ4_ACCELERATION
+        )
+        if compressed_sz > 0:
+            return <char [:compressed_sz]> self.buf
+
+        raise ValueError(f"Could not compress (error {compressed_sz})")
+
+    def decompress_col(self, str col, bytes data):
+        cdef void *mem
+        cdef size_t size
+        cdef bytes colkey_b = col.encode()
+        cdef const char *colkey_c = colkey_b
+        with nogil:
+            mem = dataset.dset_get(self.data._handle, colkey_c)
+            size = dataset.dset_getsz(self.data._handle, colkey_c)
+        if mem == NULL:
+            raise ValueError(f"Invalid column {col}")
+        else:
+            return self.decompress(data, <size_t> mem, size)
+
+    def decompress_numpy(self, bytes data, arr):
+        cdef size_t dstptr = arr.ctypes.data
+        cdef int size = arr.size
+        cdef int itemsize = arr.itemsize
+        return self.decompress(data, dstptr, size * itemsize)
+
+    def decompress(self, bytes data, size_t dstptr = 0, int dstsz = 0):
+        cdef const char *compressed = data
+        cdef int compressed_sz = len(data)
+        cdef char *uncompressed = <char *> dstptr
+        if dstsz <= 0:
+            raise ValueError("Decompression buffer size must be > 0")
+
+        if dstptr == 0:
+            self._ensure_buf(dstsz)
+            uncompressed = self.buf
+
+        cdef int uncompressed_sz = lz4.LZ4_decompress_safe(compressed, uncompressed, compressed_sz, dstsz)
+        if uncompressed_sz >= 0:
+            return <char [:uncompressed_sz]> uncompressed
+
+        raise ValueError(f"Could not decompress (error {uncompressed_sz})")