+++ cryosparc/project.py
@@ -6,7 +6,7 @@ from .workspace import Workspace
 from .job import Job, ExternalJob
 from .dataset import Dataset, DEFAULT_FORMAT
 from .row import R
-from .spec import MongoController, Datafield, Datatype, ProjectDocument
+from .spec import MongoController, Datatype, ProjectDocument, SlotSpec
 
 if TYPE_CHECKING:
     from numpy.typing import NDArray  # type: ignore
@@ -199,7 +199,7 @@ class Project(MongoController[ProjectDocument]):
         dataset: Dataset[R],
         type: Datatype,
         name: Optional[str] = None,
-        slots: Optional[List[Union[str, Datafield]]] = None,
+        slots: Optional[List[SlotSpec]] = None,
         passthrough: Optional[Tuple[str, str]] = None,
         title: Optional[str] = None,
         desc: Optional[str] = None,
@@ -260,7 +260,7 @@ class Project(MongoController[ProjectDocument]):
             type (Datatype): Type of output dataset.
             name (str, optional): Name of output on created External job. Same
                 as type if unspecified. Defaults to None.
-            slots (list[str | Datafield], optional): List of slots expected to
+            slots (list[SlotSpec], optional): List of slots expected to
                 be created for this output such as ``location`` or ``blob``. Do
                 not specify any slots that were passed through from an input
                 unless those slots are modified in the output. Defaults to None.
@@ -359,22 +359,38 @@ class Project(MongoController[ProjectDocument]):
         """
         return self.cs.download_mrc(self.uid, path_rel)
 
-    def upload(self, target_path_rel: Union[str, PurePosixPath], source: Union[str, bytes, PurePath, IO]):
+    def upload(
+        self,
+        target_path_rel: Union[str, PurePosixPath],
+        source: Union[str, bytes, PurePath, IO],
+        *,
+        overwrite: bool = False,
+    ):
         """
         Upload the given file to the project directory at the given relative
-        path.
+        path. Fails if target already exists.
 
         Args:
             target_path_rel (str | Path): Relative target path in project
                 directory.
             source (str | bytes | Path | IO): Local path or file handle to
                 upload. May also specified as raw bytes.
+            overwrite (bool, optional): If True, overwrite existing files.
+                Defaults to False.
         """
-        return self.cs.upload(self.uid, target_path_rel, source)
+        return self.cs.upload(self.uid, target_path_rel, source, overwrite=overwrite)
 
-    def upload_dataset(self, target_path_rel: Union[str, PurePosixPath], dset: Dataset, format: int = DEFAULT_FORMAT):
+    def upload_dataset(
+        self,
+        target_path_rel: Union[str, PurePosixPath],
+        dset: Dataset,
+        *,
+        format: int = DEFAULT_FORMAT,
+        overwrite: bool = False,
+    ):
         """
-        Upload a dataset as a CS file into the project directory.
+        Upload a dataset as a CS file into the project directory. Fails if
+        target already exists.
 
         Args:
             target_path_rel (str | Path): relative path to save dataset in
@@ -382,21 +398,32 @@ class Project(MongoController[ProjectDocument]):
             dset (Dataset): dataset to save.
             format (int): format to save in from ``cryosparc.dataset.*_FORMAT``,
                 defaults to NUMPY_FORMAT)
-
+            overwrite (bool, optional): If True, overwrite existing files.
+                Defaults to False.
         """
-        return self.cs.upload_dataset(self.uid, target_path_rel, dset, format=format)
+        return self.cs.upload_dataset(self.uid, target_path_rel, dset, format=format, overwrite=overwrite)
 
-    def upload_mrc(self, target_path_rel: Union[str, PurePosixPath], data: "NDArray", psize: float):
+    def upload_mrc(
+        self,
+        target_path_rel: Union[str, PurePosixPath],
+        data: "NDArray",
+        psize: float,
+        *,
+        overwrite: bool = False,
+    ):
         """
-        Upload a numpy 2D or 3D array to the job directory as an MRC file.
+        Upload a numpy 2D or 3D array to the job directory as an MRC file. Fails
+        if target already exists.
 
         Args:
             target_path_rel (str | Path): filename or relative path. Should have
                 ``.mrc`` extension.
             data (NDArray): Numpy array with MRC file data.
             psize (float): Pixel size to include in MRC header.
+            overwrite (bool, optional): If True, overwrite existing files.
+                Defaults to False.
         """
-        return self.cs.upload_mrc(self.uid, target_path_rel, data, psize)
+        return self.cs.upload_mrc(self.uid, target_path_rel, data, psize, overwrite=overwrite)
 
     def mkdir(
         self,